requirements.txt
asgiref
certifi==2020.6.20
dj-database-url==0.5.0
Django
django-cors-headers==3.5.0
django-postgres-copy==2.4.2
djangorestframework==3.12.2
gunicorn==20.0.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy
pandas
psycopg2
python-dateutil==2.8.1
pytz==2020.1
serpy==0.3.1
six
SQLAlchemy==1.3.20
sqlparse
whitenoise==5.2.0
wincertstore==0.2
